requirements.txt
albucore==0.0.13
albumentations==1.4.14
annotated-types==0.7.0
attrdict==2.0.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
et-xmlfile==1.1.0
eval-type-backport==0.2.0
filelock==3.15.4
flask==3.0.3
flask-cors==4.0.1
fonttools==4.53.1
fsspec==2024.6.1
h5py==3.11.0
idna==3.7
imageio==2.34.2
importlib-metadata==7.0.1
importlib-resources==6.4.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy-loader==0.4
loguru==0.7.2
lxml==5.2.2
markupsafe==2.1.3
matplotlib==3.9.1
mkl==2021.4.0
mkl-service==2.4.0
mkl_fft==1.3.8
mkl_random==1.2.4
mpmath==1.3.0
mrcfile==1.5.3
natsort==8.4.0
networkx==3.2.1
numexpr==2.8.7
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pip==24.0
psutil==6.0.0
py-cpuinfo==9.0.0
pydantic==2.8.2
pydantic-core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0post0
pytz==2024.1
pyyaml==6.0.1
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
setuptools==69.5.1
six==1.16.0
soupsieve==2.6
sympy==1.13.1
tbb==2021.13.0
threadpoolctl==3.5.0
tifffile==2024.7.24
tomli==2.0.1
# torch==2.3.1+cu118  # 
# torchvision==0.18.1+cu118  # 
tqdm==4.66.4
typing-extensions==4.12.2
ultralytics==8.2.62
# ultralytics-thop==2.0.0
urllib3==2.2.2
zipp==3.17.0
# cupy-cuda118  # 
